ntools_mctal

Struct Tally

pub struct Tally {

    pub id: u32,
    pub particles: Vec<Particle>,
    pub kind: TallyKind,
    pub modifier: Modifier,
    pub comment: String,
    pub region_bins: BinData,
    pub flagged_bins: BinData,
    pub user_bins: BinData,
    pub multiplier_bins: BinData,
    pub segment_bins: BinData,
    pub cosine_bins: BinData,
    pub energy_bins: BinData,
    pub time_bins: BinData,
    pub results: Vec<TallyResult>,
    pub tfc: Tfc,

}

Standard tally type data

Contains the results for any standard F tally.

Tally results are stored as a vector of TallyResult, which contains a value and corresponding relative error.

The tally fluctuation chart over all dump cycles is stored as a [TallyFluctuation].

Supported tallies

Mnemonic	Tally Description	Fn units	*Fn units

F1:P	Integrated surface current	particles	MeV

F2:P	Surface flux	particles/cm2	MeV/cm2

F4:P	Cell flux	particles/cm2	MeV/cm2

F5a:P	Flux at a point or ring detector	particles/cm2	MeV/cm2

FIP5:P	Point detectors, pinhole	particles/cm2	MeV/cm2

FIR5:P	Point detectors, planar radiograph	particles/cm2	MeV/cm2

FIC5:P	Point detectors, cyl radiograph	particles/cm2	MeV/cm2

F6:P	Cell energy deposition	MeV/g	jerks/g

+F6	Collision heating	MeV/g	N/A

F7:P	Cell fission energy deposition	MeV/g	jerks/g

F8:P	Pulse height tally	pulses	MeV

+F8:P	Charge deposition	charge	N/A

Fields
id: u32

Tally number

particles: Vec<Particle>

Particle type

kind: TallyKind

Tally type (pinhole, ring, detector, etc…)

modifier: Modifier

Tally modifier (none, *, +)

comment: String

Tally comment

region_bins: BinData

List of cell, surface, or detector numbers

flagged_bins: BinData

Number of flagged bins (total vs. direct or flagged vs. unflagged)

user_bins: BinData

Number of user bins

multiplier_bins: BinData

Number of multiplier bins

segment_bins: BinData

List of bin bounds for a radiograph tally

cosine_bins: BinData

List of the cosine values

energy_bins: BinData

List of the energy values (MeV)

time_bins: BinData

List of the time values (shakes)

results: Vec<TallyResult>

All tally values read directly

tfc: Tfc

Tally fluctuation chart data

Implementations

impl Tally

pub fn iter(&self) -> TallyIterator<'_>

Iterator over results by region (cell, surface, detector)

pub fn n_expected_results(&self) -> usize

Calculate expected number of results from MCTAl bin records

pub fn find_result<T>(&self, region: T) -> Option<&[TallyResult]>
where T: Into<f64> + Copy,

Try to find results for a region (cell, surface, detector)