ntools_mctal

Struct Tmesh

pub struct Tmesh {

    pub id: u32,
    pub particles: Vec<Particle>,
    pub geometry: Geometry,
    pub n_cora: usize,
    pub n_corb: usize,
    pub n_corc: usize,
    pub cora: Vec<f64>,
    pub corb: Vec<f64>,
    pub corc: Vec<f64>,
    pub n_voxels: usize,
    pub n_region_bins: usize,
    pub n_flagged_bins: usize,
    pub n_user_bins: usize,
    pub n_segment_bins: usize,
    pub n_multiplier_bins: usize,
    pub n_cosine_bins: usize,
    pub n_energy_bins: usize,
    pub n_time_bins: usize,
    pub results: Vec<TallyResult>,

}

TMESH tally type data

Overview

MCNP6 offers two mesh tallies. The TMESH tally was developed for MCNPX, and FMESH for MCNP5. These are very similar and both were kept when merging MCNP5 and MCNPX into MCNP6.

For whatever reason, the TMESH data are also written to the MCTAL file. These are also known in user manuals as โSuperimposed Mesh Tally Type Aโ.

Supported types

For reference, below are the supported types:

Type	Description

1	Track-Averaged Mesh Tally

2	Source Mesh Tally

3	Energy Deposition Mesh Tally

4	DXTRAN Mesh Tally

Fields
id: u32

Tally number

particles: Vec<Particle>

Particle type

geometry: Geometry

Tally type (pinhole, ring, detector, etcโฆ)

n_cora: usize

Number of i bins (i.e. fmesh iints)

n_corb: usize

Number of j bins (i.e. fmesh jints)

n_corc: usize

Number of k bins (i.e. fmesh kints)

cora: Vec<f64>

Bin edges for i coordinate (i.e. fmesh imesh)

corb: Vec<f64>

Bin edges for j coordinate (i.e. fmesh jmesh)

corc: Vec<f64>

Bin edges for k coordinate (i.e. fmesh kmesh)

n_voxels: usize

Number of voxels in the mesh

n_region_bins: usize

Number of cell or surface numbers

n_flagged_bins: usize

Number of flagged bins (total vs. direct or flagged vs. unflagged)

n_user_bins: usize

Number and type of user bins

n_segment_bins: usize

Number of segment bins

n_multiplier_bins: usize

Number of multiplier bins

n_cosine_bins: usize

Number and type of multiplier bins

n_energy_bins: usize

Number and type of energy bins

n_time_bins: usize

Number and type of time bins

results: Vec<TallyResult>

All tally values read directly

Implementations

impl Tmesh

pub fn n_expected_results(&self) -> usize

Calculate expected number of results from MCTAL bin records